Arreglo para que todo este en orden
Puse en el lugar correcto cada juego
