+++ app/src/main/java/com/example/appfinal/screens/jugar/juegos/Juego1.kt
@@ -40,10 +40,7 @@ import androidx.compose.ui.unit.sp
 fun Juego1 (navController: NavHostController){
     // Establecer el fondo azul claro
     val azulClaro = Color(173, 216, 230)
-    var images by remember { mutableStateOf(generateImages()) }
-    var visibleImages by remember { mutableStateOf(images.filter { it.isVisible }) }
-    var deletedImages by remember { mutableStateOf(mutableListOf<DraggableImage>()) }
-    var deletedImageCount by remember { mutableStateOf(0) }
+
 
     Box(
         modifier = Modifier
@@ -77,111 +74,6 @@ fun Juego1 (navController: NavHostController){
             )
         }
 
-        Box(
-            modifier = Modifier.fillMaxSize(),
-        ) {
-            Text(
-                text = "Número de imágenes visibles: ${visibleImages.size}",
-                modifier = Modifier
-                    .align(Alignment.TopStart) // Mantenido como Alignment.TopStart
-                    .padding(16.dp)
-                    .then(Modifier.fillMaxWidth()),
-                fontSize = 20.sp
-            )
-
-            Text(
-                text = "Número de imágenes eliminadas: $deletedImageCount",
-                modifier = Modifier
-                    .align(Alignment.TopStart) // Mantenido como Alignment.TopStart
-                    .padding(top = 48.dp, start = 16.dp)
-                    .then(Modifier.fillMaxWidth()),
-                fontSize = 20.sp
-            )
-
-            if (visibleImages.isNotEmpty()) {
-                visibleImages.forEach { image ->
-                    DraggableImage(image = image) {
-                        deletedImages.add(image)
-                        deletedImageCount++
-                        visibleImages = visibleImages.filter { it != image }
-                    }
-                }
-            } else {
-                val randomImageCount = Random.nextInt(1, 11) // Genera un número aleatorio entre 1 y 20
-                addNewImages(images, randomImageCount)
-                visibleImages = images
-            }
         }
-    }
-}
 
-
-
-data class DraggableImage(
-    val id: Int,
-    var offset: IntOffset,
-    val color: Color,
-    val radius: Int,
-    var isVisible: Boolean = true
-)
-
-fun addNewImages(images: MutableList<DraggableImage>, imageCount: Int) {
-    images.clear()
-
-    for (id in 1..imageCount) {
-        val xOffset = Random.nextInt(100, 1500)
-        val yOffset = Random.nextInt(100, 1000)
-        val color = Color(Random.nextFloat(), Random.nextFloat(), Random.nextFloat(), 1f)
-        val radius = Random.nextInt(150, 200) // Radio aleatorio para los círculos
-        val isVisible = true
-
-        images.add(
-            DraggableImage(
-                id = images.size + 1,
-                offset = IntOffset(xOffset, yOffset),
-                color = color,
-                radius = radius,
-                isVisible = isVisible
-            )
-        )
-    }
-}
-
-@Composable
-fun DraggableImage(image: DraggableImage, onDeleteClick: () -> Unit) {
-    Box(
-        modifier = Modifier
-            .offset { image.offset }
-            .size(image.radius.dp)
-            .fillMaxSize()
-            .background(color = image.color, shape = CircleShape)
-            .clickable {
-                image.isVisible = !image.isVisible
-                onDeleteClick()
-            }
-    ) {
-        // Agrega el contenido de la imagen aquí, si es necesario
-    }
-}
-
-fun generateImages(): MutableList<DraggableImage> {
-    val images = mutableListOf<DraggableImage>()
-
-    for (id in 1..10) {
-        val xOffset = Random.nextInt(100, 2000)
-        val yOffset = Random.nextInt(100, 1000)
-        val color = Color(Random.nextFloat(), Random.nextFloat(), Random.nextFloat(), 1f)
-        val radius = Random.nextInt(150, 200)
-
-        images.add(
-            DraggableImage(
-                id = id,
-                offset = IntOffset(xOffset, yOffset),
-                color = color,
-                radius = radius
-            )
-        )
-    }
-
-    return images
 }
